+++ net/minecraft/block/EndGatewayBlock.java
@@ -42,12 +42,12 @@ extends BlockWithEntity {
         }
         int i = ((EndGatewayBlockEntity)lv).getDrawnSidesCount();
         for (int j = 0; j < i; ++j) {
-            double d = (double)arg3.getX() + (double)random.nextFloat();
-            double e = (double)arg3.getY() + (double)random.nextFloat();
-            double f = (double)arg3.getZ() + (double)random.nextFloat();
-            double g = ((double)random.nextFloat() - 0.5) * 0.5;
-            double h = ((double)random.nextFloat() - 0.5) * 0.5;
-            double k = ((double)random.nextFloat() - 0.5) * 0.5;
+            double d = (double)arg3.getX() + random.nextDouble();
+            double e = (double)arg3.getY() + random.nextDouble();
+            double f = (double)arg3.getZ() + random.nextDouble();
+            double g = (random.nextDouble() - 0.5) * 0.5;
+            double h = (random.nextDouble() - 0.5) * 0.5;
+            double k = (random.nextDouble() - 0.5) * 0.5;
             int l = random.nextInt(2) * 2 - 1;
             if (random.nextBoolean()) {
                 f = (double)arg3.getZ() + 0.5 + 0.25 * (double)l;
